feat(user): add account deletion confirmation email
This commit introduces a new feature where a confirmation email is sent
to the user upon successful account deletion. The email content is
defined in `lib/emails.ts`. The email sending logic is implemented in
`server/api/routers/user.ts`. Additionally, some changes are made to
the user edit page and a new package '@radix-ui/react-alert-dialog' is
added to the dependencies.